# openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/attributes/{attributeId}": {
			"get": {
				"operationId": "Attributes_getById",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeDto"
								}
							}
						}
					},
					"404": {
						"description": "ATTRIBUTE_NOT_FOUND"
					}
				},
				"tags": [
					"Attributes"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"patch": {
				"operationId": "Attributes_patch",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAttributeDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeDto"
								}
							}
						}
					}
				},
				"tags": [
					"Attributes"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "Attributes_update",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAttributeDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeDto"
								}
							}
						}
					}
				},
				"tags": [
					"Attributes"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "Attributes_remove",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAttributeDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeDto"
								}
							}
						}
					}
				},
				"tags": [
					"Attributes"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/attributes/{attributeId}/values": {
			"post": {
				"operationId": "AttributeValues_save",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAttributeValueDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeValueDto"
								}
							}
						}
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"get": {
				"operationId": "AttributeValues_getValuesOfAttributeById",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AttributeValueDto"
									}
								}
							}
						}
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/attributes/{attributeId}/values/{valueId}": {
			"get": {
				"operationId": "AttributeValues_getValueById",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "valueId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeValueDto"
								}
							}
						}
					},
					"404": {
						"description": "ATTRIBUTE_VALUE_NOT_FOUND"
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"patch": {
				"operationId": "AttributeValues_patchValue",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "valueId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAttributeValueDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeValueDto"
								}
							}
						}
					},
					"404": {
						"description": "ATTRIBUTE_VALUE_NOT_FOUND"
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "AttributeValues_updateValue",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "valueId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAttributeValueDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeValueDto"
								}
							}
						}
					},
					"404": {
						"description": "ATTRIBUTE_VALUE_NOT_FOUND"
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "AttributeValues_deleteValueByID",
				"parameters": [
					{
						"name": "attributeId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "valueId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttributeValueDto"
								}
							}
						}
					}
				},
				"tags": [
					"Attribute Values"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
	},
	"info": {
		"title": "Nima CMS",
		"description": "The core api for the Nima CMS backend.",
		"version": "0.0.1-alpha",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "http://localhost:3333",
			"description": "Localhost"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"TranslatableDto": {
				"type": "object",
				"properties": {
					"de": {
						"type": "string"
					},
					"it": {
						"type": "string"
					},
					"fr": {
						"type": "string"
					},
					"es": {
						"type": "string"
					},
					"el": {
						"type": "string"
					},
					"en": {
						"type": "string"
					}
				}
			},
			"InputType": {
				"type": "string",
				"enum": [
					"DROPDOWN",
					"MULTISELECT",
					"FILE",
					"REFERENCE",
					"NUMERIC",
					"RICH_TEXT",
					"SWATCH",
					"BOOLEAN",
					"DATE",
					"DATE_TIME"
				]
			},
			"Unit": {
				"type": "string",
				"enum": [
					"CM",
					"M",
					"KM",
					"FT",
					"YD",
					"INCH",
					"SQ_CM",
					"SQ_M",
					"SQ_KM",
					"SQ_FT",
					"SQ_YD",
					"SQ_INCH",
					"CUBIC_MILLIMETER",
					"CUBIC_CENTIMETER",
					"CUBIC_DECIMETER",
					"CUBIC_METER",
					"LITER",
					"CUBIC_FOOT",
					"CUBIC_INCH",
					"CUBIC_YARD",
					"QT",
					"PINT",
					"FL_OZ",
					"ACRE_IN",
					"ACRE_FT",
					"G",
					"LB",
					"OZ",
					"KG",
					"TONNE"
				]
			},
			"CreateAttributeDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Attribute Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"availableInGrid": {
						"type": "boolean",
						"example": true
					},
					"visibleInStorefront": {
						"type": "boolean",
						"example": true
					},
					"filterableInDashboard": {
						"type": "boolean",
						"example": true
					},
					"filterableInStorefront": {
						"type": "boolean",
						"example": true
					},
					"valueRequired": {
						"type": "boolean",
						"example": true
					},
					"storefrontSearchPosition": {
						"type": "number",
						"example": 0
					},
					"inputType": {
						"example": "DROPDOWN",
						"$ref": "#/components/schemas/InputType"
					},
					"unit": {
						"$ref": "#/components/schemas/Unit"
					}
				},
				"required": [
					"name",
					"slug",
					"metadata",
					"privateMetadata",
					"availableInGrid",
					"visibleInStorefront",
					"filterableInDashboard",
					"filterableInStorefront",
					"valueRequired",
					"storefrontSearchPosition",
					"inputType"
				]
			},
			"AttributeDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number"
					},
					"name": {
						"example": {
							"en": "Attribute Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"availableInGrid": {
						"type": "boolean",
						"example": true
					},
					"visibleInStorefront": {
						"type": "boolean",
						"example": true
					},
					"filterableInDashboard": {
						"type": "boolean",
						"example": true
					},
					"filterableInStorefront": {
						"type": "boolean",
						"example": true
					},
					"valueRequired": {
						"type": "boolean",
						"example": true
					},
					"storefrontSearchPosition": {
						"type": "number",
						"example": 0
					},
					"inputType": {
						"example": "DROPDOWN",
						"$ref": "#/components/schemas/InputType"
					},
					"unit": {
						"$ref": "#/components/schemas/Unit"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"metadata",
					"privateMetadata",
					"availableInGrid",
					"visibleInStorefront",
					"filterableInDashboard",
					"filterableInStorefront",
					"valueRequired",
					"storefrontSearchPosition",
					"inputType"
				]
			},
			"UpdateAttributeDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Attribute Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"availableInGrid": {
						"type": "boolean",
						"example": true
					},
					"visibleInStorefront": {
						"type": "boolean",
						"example": true
					},
					"filterableInDashboard": {
						"type": "boolean",
						"example": true
					},
					"filterableInStorefront": {
						"type": "boolean",
						"example": true
					},
					"valueRequired": {
						"type": "boolean",
						"example": true
					},
					"storefrontSearchPosition": {
						"type": "number",
						"example": 0
					},
					"inputType": {
						"example": "DROPDOWN",
						"$ref": "#/components/schemas/InputType"
					},
					"unit": {
						"$ref": "#/components/schemas/Unit"
					}
				}
			},
			"CreateAttributeValueDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "green"
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"value": {
						"type": "string",
						"example": "Green"
					},
					"fileUrl": {
						"type": "string"
					},
					"richText": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"boolean": {
						"type": "boolean"
					},
					"dateTime": {
						"type": "string"
					}
				},
				"required": [
					"name"
				]
			},
			"AttributeValueDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "green"
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"value": {
						"type": "string",
						"example": "Green"
					},
					"fileUrl": {
						"type": "string"
					},
					"richText": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"boolean": {
						"type": "boolean"
					},
					"dateTime": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				]
			},
			"UpdateAttributeValueDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "green"
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"value": {
						"type": "string",
						"example": "Green"
					},
					"fileUrl": {
						"type": "string"
					},
					"richText": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"boolean": {
						"type": "boolean"
					},
					"dateTime": {
						"type": "string"
					}
				}
			},
...
		}
	}
}